requirements.txt
langchain==0.3.27
langgraph==0.6.7
langchain-aws==0.2.18
langchain-core==0.3.76
tiktoken==0.11.0
fastapi>=0.109.1,<1
requests==2.32.4
types-requests==2.31.0
pydantic==2.10.2
pyhumps==3.8.0
uvicorn>=0.23.1,<1
python-ulid==1.1.0
boto3==1.37.0
aws-xray-sdk==2.14.0
python-jose==3.4.0
pg8000==1.30.3
argparse==1.4.0
tenacity<=8.3.0
langdetect==1.0.9
types-retry>=0.9.9.4,<1
opensearch-py>=2.0.0
requests-aws4auth>=1.0.0
duckduckgo-search==7.3.0
boto3-stubs[bedrock,bedrock-agent-runtime,bedrock-runtime,boto3]==1.37.0
# firecrawl-py==1.11.1
reretry==0.11.8
langfuse==3.3.4
pytz==2025.2